+++ xmonad-manage/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## master
 - Added `remove` command, which removes specified profile.
+- Fixed `install` not creating required directories.
 
 ## 0.4
 (Changelog from this point is lost)